Much improved arsenal.
Create a new building with ammo crates. The building starts out locked.

The ACE defuse kit is used as a stand-in prop. Once the BLUFOR has been
detected, anyone with that kit can unlock or lock the front door.

Added some lights because otherwise nobody can see inside the building.